Refactor conjecture class in ce guided instantiation, move to own file.  In preparation for sygus streaming mode.
