+++ src/options/quantifiers_options
@@ -287,6 +287,9 @@ option sygusCRefEval --sygus-cref-eval bool :default true
 option sygusCRefEvalMinExp --sygus-cref-eval-min-exp bool :default true
   use min explain for direct evaluation of refinement lemmas for conflict analysis
   
+option sygusStream --sygus-stream bool :default false
+  enumerate a stream of solutions instead of terminating after the first one
+  
 # approach applied to general quantified formulas
 option cbqi --cbqi bool :read-write :default false
  turns on counterexample-based quantifier instantiation
